Fix token generation when URL contains encoded characters
- I got 403 error when I executed following query /myIndex/entity1,entity2/_search?...
After debugging and reading manuals/error messages I figured out what was happening.
If URL contains encoded characters they need to be double encoded, so URL in form
"test%2Ctest" should be transformed to "test%252Ctest". The whole URL can't be
encoded because it will lead to / character replace, hence first it needs to be split to
parts and then each part should be encoded (double encoded in fact). After that
URL should be joined together